fix keyboard android
