+++ app.json
@@ -23,7 +23,6 @@
         "backgroundColor": "#ffffff"
       },
       "edgeToEdgeEnabled": true,
-      "softwareKeyboardLayoutMode": "pan",
       "package": "com.anonymous.bitnovoTecnica"
     },
     "web": {
