+++ package.json
@@ -25,6 +25,7 @@
     "react-native": "0.79.2",
     "react-native-currency-input": "^1.1.1",
     "react-native-gesture-handler": "~2.24.0",
+    "react-native-keyboard-accessory": "^0.1.16",
     "react-native-qrcode-svg": "^6.3.15",
     "react-native-reanimated": "~3.17.4",
     "react-native-safe-area-context": "5.4.0",
